Renamed plugin to ClansPE (Official)
This has been renamed. WARNING TO ALL:
If you're going to install this plugin, keep in mind that the factions originally there could potentially get deleted. IF it doesn't, then you're lucky. BUt the chances are most likely that your factions will be deleted. If you want to continue using the old versions of this plugin, please use the old FactionsPro plugin. We are going to make the Branch for Pocketmine users only, but it will contain the old facts, the good ol faction plugin. For you, and for the awesome people out in the community.
If you however want this new plugin, then go ahead and install. Some more features are coming on the way soon. Enjoy.